Refactored headers according to recommendations in discussion on Discord
