--- src/common.hpp
@@ -1,16 +1,0 @@
-#pragma once
-
-#include "apiClient.hpp"
-#include "httplib.h"
-#include "json.hpp"
-#include "matchBuilder.hpp"
-#include "utilities.hpp"
-#include "threadWorks.hpp"
-#include "mapping.hpp"
-#include <csignal>
-#include <chrono>
-#include <thread>
-#include <vector>
-#include <mutex>
-#include <atomic>
-#include <iostream>
