+++ src/millisecondClock.hpp
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <chrono>
+
+class MillisecondClock {
+public:
+    // Start the clock by recording the current time
+    void start();
+
+    // Get the elapsed time in milliseconds since the clock was started
+    long long perSecondCheck();
+
+    // Get the elapsed time in milliseconds since the clock was started
+    long long elapsedMilliseconds();
+
+    // Reset the clock to the current time
+    void resetClock();
+
+private:
+    std::chrono::high_resolution_clock::time_point initialStart; // Store the start time
+    std::chrono::high_resolution_clock::time_point currentStart;
+};
