+++ src/myRandom.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <string>
+
+class myRandom {
+public:
+    // Generates a random integer between min and max (inclusive)
+    static int generateRandomInt(int min, int max);
+
+    // Generates a random string of a specified length containing letters and digits
+    static std::string generateRandomString(size_t length);
+
+    // Generates a random numeric string of a specified length
+    static std::string generateRandomNumberString(size_t length);
+
+    // Generates a random boolean (true or false)
+    static bool getRandomBool();
+};
