+++ src/oceanBuilder.hpp
@@ -1,18 +1,14 @@
 #pragma once
-#include <iostream>
-#include <fstream>
-#include <string>
-#include <random>
-#include "json.hpp"
-#include "utilities.hpp"
-#include "mapping.hpp"
 
-using json = nlohmann::json;
+#include <string>
+
+// project dependencies
+#include "json.hpp"
 
 class oceanBuilder {
 public:
     // Function to generate a random match based on the template
-    json randomOcean();
+    nlohmann::json randomOcean();
 
 private:
     // Internal function to remove first and last characters from a string
@@ -21,4 +17,3 @@ private:
     // Internal function to get a random key from a JSON file
     static std::string getRandomFromJson();
 };
-
